release: v0.3.2 - real-VOD benchmark + frame-loss and CLI robustness fixes
Benchmark on two real Twitch VODs (1h28m 1080p60, 6h26m 720p60) exposed two critical frame-loss defects and a set of CLI bugs on long content; all fixed (docs/benchmarks.md).

- P0 frame-hole gate: resume/tree/final-output gates now reject video whose decoded frame count is far below duration x nominal fps (catches PTS-hole corruption duration checks miss)

- P1 UTF-8 stdio: reconfigure stdout/stderr for the whole CLI run so piped/redirected output on cp1251/cp866 no longer dies with UnicodeEncodeError

- P1 isatty guard: software-fallback consent + legacy-rename prompts refuse instead of blocking forever on non-interactive stdin

- P2 MemoryMonitor: RSS now sums the whole process tree (shim/launcher no longer hides the real encode)

- batch: split keep blocks >300s into contiguous sub-trims (ffmpeg 9.x concat-filter frame-loss workaround)

- segment/batch/cut/audio: encode+validate timeout scales with part length (flat 600s cap no longer kills multi-hour encodes)

+++ stream2video/concat/runner.py
@@ -46,7 +46,7 @@ logger = logging.getLogger(__name__)
 def _run_ffmpeg(
     cmd: list[str],
     progress_callback: Callable[[float], None] | None,
-    timeout: int,
+    timeout: float,
     label: str = "ffmpeg",
     cancel_callback: Callable[[], bool] | None = None,
     track_progress: bool = True,
@@ -497,7 +497,7 @@ def _run_ffmpeg(
 
 def _wait_with_cancel(
     process: subprocess.Popen,
-    timeout: int,
+    timeout: float,
     cancel_callback: Callable[[], bool] | None,
     label: str,
 ) -> int:
